starting regression prediction

+++ ams/pipes/batchy_bae.py
@@ -31,48 +31,81 @@ def revert_files_to_original_location(dest_dir_path: Path, files_in_transition: 
         shutil.move(str(f), str(dest_path))
 
 
-def remove_remaining_files(target_path: Path) -> Path:
-    files = file_services.list_files(target_path, use_dir_recursion=False)
+def delete_handle_error(dir_path: Path):
+    """
+    If the error is due to an access error (read only file)
+    it attempts to add write permission and then retries.
 
-    tmp_trash = file_services.create_unique_folder_name(constants.TWITTER_TRASH_OUTPUT, prefix="batchy_bae")
+    If the error is for another reason it re-raises the error.
+
+    Usage : ``shutil.rmdir(path, onerror=onerror)``
+    """
+    file_services.remove_folder_read_only(dir_path=dir_path, recursive=True)
+
+    files = file_services.list_files(dir_path, use_dir_recursion=True)
     for f in files:
-        file_path_new = Path(tmp_trash, f.name)
-        shutil.move(str(f), str(file_path_new))
+        f.unlink()
 
+    folders = file_services.list_child_folders(dir_path)
+    for fld in folders:
+        delete_handle_error(fld)
+
+    try:
+        dir_path.rmdir()
+    # NOTE: 2021-02-27: chris.flesche: May have already been deleted.
+    except FileNotFoundError as fe:
+        pass
+
+
+def remove_remaining_files(target_path: Path, should_delete: bool):
+    files = file_services.list_files(target_path, use_dir_recursion=True)
     folders = file_services.list_child_folders(target_path)
-    for f in folders:
-        file_path_new = Path(tmp_trash, f.name)
-        shutil.move(str(f), str(file_path_new))
 
-    return Path(tmp_trash)
+    if should_delete:
+        for f in files:
+            f.unlink()
+        for fld in folders:
+            delete_handle_error(fld)
+    else:
+        tmp_trash = file_services.create_unique_folder_name(constants.TWITTER_TRASH_OUTPUT, prefix="batchy_bae")
+        for f in files:
+            file_path_new = Path(tmp_trash, f.name)
+            if f.exists():
+                shutil.move(str(f), str(file_path_new))
+
+        for f in folders:
+            file_path_new = Path(tmp_trash, f.name)
+            if f.exists():
+                shutil.move(str(f), str(file_path_new))
 
 
-def get_stage_dir(source_dir_path: Path):
-    return Path(source_dir_path.parent, STAGING_FOLDER_NAME)
+def get_stage_dir(src_dir_path: Path):
+    return Path(src_dir_path.parent, STAGING_FOLDER_NAME)
 
 
-def revert_in_transition_files(source_dir_path: Path):
-    staging_path = get_stage_dir(source_dir_path=source_dir_path)
+def revert_in_transition_files(src_dir_path: Path):
+    staging_path = get_stage_dir(src_dir_path=src_dir_path)
     ensure_dir(staging_path)
     files_in_transition = get_files_in_transition(staging_path)
 
-    revert_files_to_original_location(dest_dir_path=source_dir_path, files_in_transition=files_in_transition)
+    revert_files_to_original_location(dest_dir_path=src_dir_path, files_in_transition=files_in_transition)
 
-    remove_remaining_files(staging_path)
+    remove_remaining_files(staging_path, should_delete=True)
 
 
-def start(source_path: Path, output_dir_path: Path, process_callback: pipe_process_callback, should_archive: bool = True, snow_plow_stage: bool = True):
+def start(source_path: Path, out_dir_path: Path, process_callback: pipe_process_callback, should_delete_leftovers: bool, should_archive: bool = True,
+          snow_plow_stage: bool = True):
     if snow_plow_stage:
-        target_path = get_stage_dir(source_dir_path=source_path)
-        remove_remaining_files(target_path=target_path)
+        target_path = get_stage_dir(src_dir_path=source_path)
+        remove_remaining_files(target_path=target_path, should_delete=should_delete_leftovers)
     else:
         # NOTE: 2021-02-07: chris.flesche: Helps with debugging.
-        revert_in_transition_files(source_dir_path=source_path)
+        revert_in_transition_files(src_dir_path=source_path)
 
     if file_services.has_no_files(the_path=source_path):
         return False
     else:
-        process(source_path, output_dir_path=output_dir_path, process_callback=process_callback, should_archive=should_archive)
+        process(source_path, out_dir_path=out_dir_path, process_callback=process_callback, should_archive=should_archive)
         return True
 
 
@@ -108,42 +141,41 @@ def archive(source_path: Path, staging_dir_path: Path):
         logger.info("WARNING: No files found to archive.")
 
 
-def unstage(source_path: Path, output_dir_path: Path):
+def unstage(source_path: Path, out_dir_path: Path):
     logger.info("Unstaging...")
     files_in_transition = get_files_in_transition(source_path)
     total_files = len(files_in_transition)
     if total_files > 0:
         for f in files_in_transition:
             filename = f.name
-            dest_path = Path(output_dir_path, filename[:-len(IN_TRANSITION_ENDING)])
+            dest_path = Path(out_dir_path, filename[:-len(IN_TRANSITION_ENDING)])
             shutil.move(str(f), str(dest_path))
 
     if total_files == 0:
         logger.info("WARNING: No files found to archive.")
 
 
-def ensure_clean_output_path(output_dir_path: Path):
-    if not file_services.is_empty(output_dir_path):
+def ensure_clean_output_path(out_dir_path: Path, should_delete_remaining: bool):
+    if not file_services.is_empty(out_dir_path):
         logger.info("Cleaning output path ...")
-        trash_path = remove_remaining_files(output_dir_path)
-        logger.warning(f"Output folder '{output_dir_path}' is not empty. Files moved to '{trash_path}'.")
+        remove_remaining_files(out_dir_path, should_delete=should_delete_remaining)
 
 
-def process(source_path: Path, output_dir_path: Path, process_callback: Callable[[Path, Path], None], should_archive: bool = True):
+def process(source_path: Path, out_dir_path: Path, process_callback: Callable[[Path, Path], None], should_archive: bool = True):
     if not source_path.exists():
         raise Exception(f"Source folder does not exist: '{source_path}'.")
 
     staging_dir_path = move_to_staging(source_path)
     logger.info(f"Moved to staging folder '{staging_dir_path}'.")
 
-    process_callback(staging_dir_path, output_dir_path)
+    process_callback(staging_dir_path, out_dir_path)
 
     if should_archive:
         logger.info("Will archive.")
         archive(source_path=source_path, staging_dir_path=staging_dir_path)
     else:
         logger.info("Will unstage.")
-        unstage(source_path=staging_dir_path, output_dir_path=source_path)
+        unstage(source_path=staging_dir_path, out_dir_path=source_path)
 
 
 if __name__ == '__main__':
@@ -155,4 +187,4 @@ if __name__ == '__main__':
         pass
 
 
-    start(source_path=source_dir_path, output_dir_path=output_dir_path, process_callback=foo, should_archive=False)
+    start(source_path=source_dir_path, out_dir_path=output_dir_path, process_callback=foo, should_archive=False, should_delete_leftovers=False)
